requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==4.2.7
django-crontab==0.7.1
django-filter==23.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
tzdata==2023.3
